edge/requirements.txt
pyserial==3.5
Flask==2.3.3
Werkzeug==2.3.7
opencv-python-headless==4.8.1.78
numpy==1.26.2
